% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frenchfish.R
\name{areAllNonnegativeIntegers}
\alias{areAllNonnegativeIntegers}
\title{Helper function to check if all values in the input count
matrix are either NA, NaN, or non-negative integers}
\usage{
areAllNonnegativeIntegers(count_matrix)
}
\arguments{
\item{count_matrix}{The count matrix}
}
\value{
TRUE if all values in count_matrix are non-NA/NaN, non-negative 
integers; otherwise FALSE.
}
\description{
Helper function to check if all values in the input count
matrix are either NA, NaN, or non-negative integers
}
