% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frenchfish.R
\name{checkAutomaticCountsEstimatesArguments}
\alias{checkAutomaticCountsEstimatesArguments}
\title{Helper function to check the arguments input to getAutomaticCountsEstimates}
\usage{
checkAutomaticCountsEstimatesArguments(probeCounts, radius, height)
}
\arguments{
\item{probeCounts}{A dataframe where the first column contains the areas of 
the nuclear blobs (this column must be named "area" and the unit of its 
entries must be the square of the unit used to measure \code{radius} and 
\code{height}) and the remaining columns (one per probe) contain the spot 
counts for different probes in each nuclear blob}

\item{radius}{The cell radius (must be measured in same unit as 
\code{height})}

\item{height}{The section height (must be measured in same unit as 
\code{radius})
#' @return Nothing if all checks are passed; otherwise throws an error or 
warning message}
}
\description{
Helper function to check the arguments input to getAutomaticCountsEstimates
}
