% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frenchfish.R
\name{checkManualCountsEstimatesArguments}
\alias{checkManualCountsEstimatesArguments}
\title{Helper function to check the arguments input to getManualCountsEstimates}
\usage{
checkManualCountsEstimatesArguments(probeCounts, radius, height)
}
\arguments{
\item{probeCounts}{A dataframe of manual spot counts with columns for 
probes and rows for nuclei}

\item{radius}{The cell radius (must be measured in same unit as 
\code{height})}

\item{height}{The section height (must be measured in same unit as 
\code{radius})}
}
\description{
Helper function to check the arguments input to getManualCountsEstimates
}
