% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frenchfish.R
\name{convertFishalyserCsvToCountMatrix}
\alias{convertFishalyserCsvToCountMatrix}
\title{Function to convert CSV output of the FISHalyseR automatic FISH 
splot counting software to a count matrix suitable for input to 
frenchFISH's getAutomaticCountsEstimates.}
\usage{
convertFishalyserCsvToCountMatrix(pathToFishalyserCsv)
}
\arguments{
\item{pathToFishalyserCsv}{The path to the CSV file of automatic spot 
counts outputted by FISHalyseR}
}
\value{
A count matrix suitable for input to getAutomaticCountsEstimates
}
\description{
Function to convert CSV output of the FISHalyseR automatic FISH 
splot counting software to a count matrix suitable for input to 
frenchFISH's getAutomaticCountsEstimates.
}
\examples{
probeCounts<-convertFishalyserCsvToCountMatrix(
    system.file("extdata", "SampleFISH.jpg_data.csv", package="frenchFISH"))
}
