% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frenchfish.R
\name{generatePPdat}
\alias{generatePPdat}
\title{Helper function to convert spot counts and nuclear area measurements into
continuous events for Poisson point estimation}
\usage{
generatePPdat(area, spots)
}
\arguments{
\item{area}{The nuclear area}

\item{spots}{The number of spots counted}
}
\value{
Vector of continuous events for Poisson point estimation
}
\description{
Helper function to convert spot counts and nuclear area measurements into
continuous events for Poisson point estimation
}
\examples{
ppDat<-generatePPdat(c(300,500,450), c(2,0,4))
}
