% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frenchfish.R
\name{getAverageVolumeFrac}
\alias{getAverageVolumeFrac}
\title{Helper function to get the average volume of nucleus sampled
given the nucleus radius and tissue section height (thickness).}
\usage{
getAverageVolumeFrac(r, h)
}
\arguments{
\item{r}{The nuclear radius}

\item{h}{The section height (thickness)}
}
\value{
The average volume of nucleus sampled given the nucleus radius and 
section height
}
\description{
Helper function to get the average volume of nucleus sampled
given the nucleus radius and tissue section height (thickness).
}
