% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frenchfish.R
\name{getMaxVolumeFrac}
\alias{getMaxVolumeFrac}
\title{Helper function to get the maximum possible volume of nucleus
sampled given the nucleus radius and section height}
\usage{
getMaxVolumeFrac(r, h)
}
\arguments{
\item{r}{The nuclear radius}

\item{h}{The section height}
}
\value{
The maximum possible volume of nucleus sampled given the nucleus 
radius and section height
}
\description{
Helper function to get the maximum possible volume of nucleus
sampled given the nucleus radius and section height
}
