% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frenchfish.R
\name{getMinVolumeFrac}
\alias{getMinVolumeFrac}
\title{Helper function that returns the minimum possible volume of nucleus
sampled given the nucleus radius and section height}
\usage{
getMinVolumeFrac(r, h)
}
\arguments{
\item{r}{The radius of the nuclei}

\item{h}{The height of the section}
}
\value{
The minimum possible volume of nucleus sampled given the nucleus
}
\description{
Helper function that returns the minimum possible volume of nucleus
sampled given the nucleus radius and section height
}
