% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frenchfish.R
\name{getPPcountsEstimates}
\alias{getPPcountsEstimates}
\title{Main frenchFISH function for generating poisson point estimates of spot counts from
counts which have been automatically generated.
Takes a data frame where the first column is nuclear area and the remaining columns
are spot counts (rows are nuclear blobs).}
\usage{
getPPcountsEstimates(countMatrix, radius, height)
}
\arguments{
\item{countMatrix}{A dataframe where the first column is nuclear area and the remaining columns are spot counts}

\item{radius}{The cell radius}

\item{height}{The section height}
}
\value{
The \code{countMatrix} with spot counts that have been adjusted using Poisson modelling
}
\description{
Main frenchFISH function for generating poisson point estimates of spot counts from
counts which have been automatically generated.
Takes a data frame where the first column is nuclear area and the remaining columns
are spot counts (rows are nuclear blobs).
}
\examples{
add(1, 1)
add(10, 1)
}
