% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frenchfish.R
\name{getVsegFrac}
\alias{getVsegFrac}
\title{Helper function that returns the fraction of the nucleus sampled for a specified
distance from the midpoint}
\usage{
getVsegFrac(d, h, r)
}
\arguments{
\item{d}{The distance sampled from the midpoint}

\item{h}{The height of the section}

\item{r}{The radius of the nuclei}
}
\value{
The fraction of the nucleus sampled for a specified distance from the midpoint
}
\description{
Helper function that returns the fraction of the nucleus sampled for a specified
distance from the midpoint
}
\examples{
vsegFrac<-getVsegFrac(5,4,8)
}
