% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frenchfish.R
\name{areAllNonnegativeIntegers}
\alias{areAllNonnegativeIntegers}
\title{Helper function to check if all values in the input count
dataframe are either NA, NaN, or non-negative integers}
\usage{
areAllNonnegativeIntegers(df)
}
\arguments{
\item{df}{The count dataframe}
}
\value{
TRUE if all values in df are non-NA/NaN, non-negative integers; 
otherwise FALSE
}
\description{
Helper function to check if all values in the input count
dataframe are either NA, NaN, or non-negative integers
}
